requirements.txt
google-cloud-api-keys
google-cloud-speech
google-cloud-texttospeech
fairscale
fire
sentencepiece
torch
black==23.9.1
lama-cpp-python